feat: add middleware parsing JWT

+++ users/models.go
@@ -7,7 +7,7 @@ import (
 // Note that, `users` table is managed by "Authentication-Server"
 
 type User struct {
-	ID         int       `json:"id" gorm:"primaryKey"`
+	ID         int64     `json:"id" gorm:"primaryKey"`
 	Email      string    `json:"-" gorm:"not null"`
 	Password   string    `json:"-" gorm:"not null"`
 	Name       string    `json:"-" gorm:"not null"`
